Remove load_path rewriting in processor
Since we supply the filename in e0d066d it's no longer required to
change to load paths in the processor. Load paths should only be used for
files unreachable from the projects for example when bundles with as a
gem or when bower is used.

This further allows us to remove the build_path, since get_files takes
care of this for us.